pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "Bootcamp blockchain training"
authors = ["UJ Blockchain"]
license = "MIT"
readme = "README.rst"

[tool.isort]
multi_line_output = 5
line_length = 119



[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.11"
python-decouple = "^3.8"
dotmap = "^1.3.30"
django-split-settings = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

